Refactor dashboard package structure
-  Everything related to redux state, event and command handling
   is now concentrated inside the 'model' directory
-  The model dir is where all the dashboard state + logic should be concentrated
-  Model is headless, there are no presentation concerns

JIRA: RAIL-3383
