+++ libs/sdk-ui-dashboard/src/dashboard/index.ts
@@ -1,16 +1,2 @@
 // (C) 2021 GoodData Corporation
 export { IDashboardProps, Dashboard } from "./Dashboard";
-export {
-    DashboardState,
-    DashboardDispatch,
-    useDashboardDispatch,
-    useDashboardSelector,
-    loadingSelector,
-    LoadingState,
-    layoutSelector,
-    LayoutState,
-    filterContextSelector,
-    FilterContextState,
-    insightsSelector,
-    DashboardContext,
-} from "./state";
