+++ libs/sdk-ui-dashboard/src/index.ts
@@ -6,8 +6,4 @@
 export * from "./dashboard";
 export * from "./filterBar";
 export * from "./topBar";
-export * from "./commands/dashboard";
-
-export { IDashboardEvent, DashboardLoaded, DashboardEventType, DashboardEvents } from "./events/dashboard";
-
-export * from "./events/eventHandler";
+export * from "./model";
