+++ libs/sdk-ui-dashboard/src/model/commandHandlers/rootCommandHandler.ts
@@ -1,6 +1,6 @@
 // (C) 2021 GoodData Corporation
 import { actionChannel, call, getContext, take } from "redux-saga/effects";
-import { DashboardCommands } from "../../commands/dashboard";
+import { DashboardCommands } from "../commands/dashboard";
 import { loadDashboardCommandHandler } from "./loadDashboardCommandHandler";
 import { DashboardContext } from "../state/dashboardStore";
