+++ libs/sdk-ui-dashboard/src/model/commands/dashboard.ts
