+++ libs/sdk-ui-dashboard/src/model/state/index.ts
