+++ libs/sdk-ui-dashboard/src/model/state/insights/index.ts
