+++ libs/sdk-ui-dashboard/src/model/state/insights/insightsEntityAdapter.ts
