+++ libs/sdk-ui-dashboard/src/model/state/insights/insightsSelectors.ts
