+++ libs/sdk-ui-dashboard/src/model/state/layout/index.ts
