+++ libs/sdk-ui-dashboard/src/model/state/layout/layoutReducers.ts
