+++ libs/sdk-ui-dashboard/src/model/state/layout/layoutSelectors.ts
