+++ libs/sdk-ui-dashboard/src/model/state/layout/layoutState.ts
